Rename network component into network-and-origin row
