+++ app/components/Views/confirmations/components/info/transfer/transfer.tsx
@@ -13,10 +13,10 @@ import { useMaxValueRefresher } from '../../../hooks/useMaxValueRefresher';
 import { useTokenAmount } from '../../../hooks/useTokenAmount';
 import { useTransferAssetType } from '../../../hooks/useTransferAssetType';
 import { HeroRow } from '../../rows/transactions/hero-row';
+import { NetworkAndOriginRow } from '../../rows/transactions/network-and-origin-row';
 import FromTo from '../../rows/transactions/from-to';
 import GasFeesDetails from '../../rows/transactions/gas-fee-details';
-import AdvancedDetailsRow from '../../rows/transactions/advanced-details-row/advanced-details-row';
-import NetworkRow from '../../rows/transactions/network-row';
+import AdvancedDetailsRow from '../../rows/transactions/advanced-details-row';
 
 const Transfer = () => {
   // Set navbar as first to prevent Android navigation flickering
@@ -44,7 +44,7 @@ const Transfer = () => {
     <View>
       <HeroRow />
       <FromTo />
-      <NetworkRow />
+      <NetworkAndOriginRow />
       {isDappTransfer && (
         <SimulationDetails
           transaction={transactionMetadata as TransactionMeta}
